+++ discovery/register.go
